Support building custom URL for prerendering.
We're needing to remove utm_* params and other query string params from the URLs we're building. Instead of overriding the request object's `build_absolute_uri` method, it makes sense to give custom backends a hook into how the URLs are constructed. The default is current behavior, so this is fully backwards compatible.